attempt at support for cdk import
import will enable destroy on a resource that has a retain policy, and then subsequent import into a new/updated project.

cls import command will be hidden until cdk import is out of preview and/or feature complete enough to support import of simple resources.

currently the error messages from cdk import are not enough to sort out support for the capability.